requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
argcomplete==3.1.4
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
bitarray==3.8.0
blinker==1.7.0
Brlapi==0.8.5
Brotli==1.1.0
certifi==2023.11.17
chardet==5.2.0
ckzg==2.1.5
click==8.1.6
cloud-init==25.2
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==41.0.7
cupshelpers==1.0
cytoolz==1.1.0
dateparser==1.2.2
dbus-python==1.3.2
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
frozenlist==1.7.0
gbinder-python==1.1.2
gyp==0.1
hexbytes==1.3.1
httplib2==0.20.4
idna==3.6
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
Mako==1.3.2.dev0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.6.4
mutagen==1.46.0
netaddr==0.8.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
packaging==24.0
parsimonious==0.10.0
pexpect==4.9.0
pillow==10.2.0
pipx==1.4.3
platformdirs==4.2.0
propcache==0.3.2
psutil==5.9.8
ptyprocess==0.7.0
PyAudio==0.2.13
pycairo==1.25.1
pycryptodome==3.23.0
pycryptodomex==3.20.0
pycups==2.0.1
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu5.2
python-binance==1.0.29
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
pytz==2025.2
pyunormalize==17.0.0
pyxdg==0.28
PyYAML==6.0.1
regex==2025.9.1
requests==2.31.0
rich==13.7.1
rlp==4.1.0
setuptools==68.1.2
six==1.16.0
systemd-python==235
toolz==1.1.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
userpath==1.9.1
wadllib==1.3.6
web3==7.14.0
websockets==15.0.1
wheel==0.42.0
xdg==5
xkit==0.0.0
yarl==1.20.1
yt-dlp==2024.4.9

# --- Project-critical dependencies for fresh deployments ---
# WS price stream (module: websocket)
websocket-client==1.9.0
# News guard calendar provider
investpy==1.0.8
# Core data/ML stack used by AI managers and training pipelines
pandas==2.2.3
scikit-learn==1.5.2
xgboost==2.1.1
lightgbm==4.5.0
catboost==1.2.7
torch==2.4.1
# C++ acceleration Python-side build/runtime requirements
pybind11==2.11.1
cmake==3.30.5

# Note: C++ native build also needs OS packages (not pip-installable):
# g++, libssl-dev, libcurl4-openssl-dev
